// apps/admin/src/lib/api/generated/models/sourcePreviewOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DevFeed Admin API
 * Private administration API. OIDC sessions and CSRF protection required.
 * OpenAPI spec version: 0.0.5
 */

export interface SourcePreviewOut {
  description?: string | null;
  entries_seen: number;
  entries_skipped: number;
  image_url?: string | null;
  language?: string | null;
  logo_url?: string | null;
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
  warnings: string[];
  website_url?: string | null;
}